pyproject.toml
[build-system]
requires = [
    "setuptools",
    "setuptools-scm",
]

build-backend = "setuptools.build_meta"

[project]
name = "bridger"
dynamic = ["version"]
description = "A bridge between MQTT and InfluxDB for Meshtastic metrics"
readme = "README.md"

authors = [
    { name="Andy Shinn", email="andys@andyshinn.as" },
]

dependencies = [
    "aiohttp",
    "aiomqtt",
    "aiocache",
    "cryptography",
    "dataclasses-json",
    "discord.py",
    "influxdb-client",
    "loguru",
    "meshcoredecoder",
    "meshtastic",
    "paho-mqtt",
    "protobuf",
    "pytest",
    "pytest-cov",
    "python-dotenv",
    "pytest-aiohttp",
    "requests",
    "requests-mock",
    "rich",
    "sentry-sdk",
    "tenacity",
]

[project.urls]
Source = "https://github.com/austinmesh/bridger"
Graphs = "https://graphs.austinmesh.org"

[tool.black]
line-length = 125

[tool.isort]
profile = "black"
line_length = 125

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
consider_namespace_packages = true
asyncio_mode = "auto"
addopts = [
    "--cov=bridger",
    "--cov-report=term-missing",
]
